\name{Internals}
\alias{gam.fit2}
\alias{gam.fit3}
\title{
  Internals function coming from the mgcv package
}
\description{
  The mgcv package does not provide the penalized deviance of a fitted
  gam. Thus, these are modified functions of the mgcv package to get
  this penalized deviance.
  }
\author{Simon Wood}
\keyword{internal}
