\name{print.joint}
\alias{print.joint}
\title{Printing joint object}
\description{
  Printing joint oject
}
\usage{
\method{print}{joint}(x, \dots)
}
\arguments{
  \item{x}{An object of class \code{"joint"}. Most often the return of
    the \code{\link{fitjoint}} function.}
  \item{\dots}{Other option to be passed to the \code{\link{plot}}
    function.}
}
\seealso{
  \code{\link{fitjoint}}
}
\author{Mathieu Ribatet and Bertrand Iooss}
\keyword{print}
