\name{summary.joint}
\alias{summary.joint}
\title{Summary for objects of class joint}
\description{
  Summary for joint objects
}
\usage{
\method{summary}{joint}(object, \dots)
}
\arguments{
  \item{object}{An object of class ``joint''. Most often this object is
    returned by the \code{\link{fitjoint}} function.}
  \item{\dots}{Other options to be passed to the \code{\link{summary}}
    function.}
}
\value{
  Several useful informations about the fitted ``joint'' object.
}
\seealso{
  \code{\link{fitjoint}}, \code{\joint{summary}}
}
\author{Mathieu Ribatet and Bertrand Iooss}
\keyword{print}
